feat(provider): add taskGetProfile, fetchProfile
