+++ src/social-network-provider/facebook.com/worker-provider.ts
@@ -1,6 +1,7 @@
 import { defineSocialNetworkWorker } from '../../social-network/worker'
 import { sharedProvider } from './shared-provider'
-import { fetchPostContentFacebook } from './fetchPostContent'
+import { fetchPostContentFacebook } from './Worker/fetchPostContent'
+import { fetchProfileFacebook } from './Worker/fetchProfile'
 
 defineSocialNetworkWorker({
     ...sharedProvider,
@@ -15,4 +16,5 @@ defineSocialNetworkWorker({
         url: [{ hostEquals: 'www.facebook.com' }, { hostEquals: 'm.facebook.com' }],
     },
     fetchPostContent: fetchPostContentFacebook,
+    fetchProfile: fetchProfileFacebook,
 })
